Break out migration store unittests to separate file
